chore: replace logrus with slog
