+++ go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/bwmarrin/snowflake v0.3.0
 	github.com/elliotwms/bot v0.3.33
 	github.com/elliotwms/fakediscord v0.18.2
-	github.com/sirupsen/logrus v1.9.3
 	github.com/stretchr/testify v1.10.0
 )
